refactor(rtems-boot): make the stats funnel a per-OS backend selection
fd_usage/mem_usage each carried their own `#[cfg]` pair, so teaching the
funnel a second OS meant one new arm per function per OS, and the extern
declarations sat under an item-level cfg that a later edit could widen
without noticing.

Split `stats.rs` into `stats/{mod,rtems,unsupported}.rs`. The funnel now
carries no `#[cfg]` at all: it delegates to a `backend` module chosen by
exactly one cfg pair at the top of `mod.rs`. A new OS becomes a new file
plus one `mod` line, and no consumer changes — consumers already see
`Option`, which means "this build has no reading".

Gating the RTEMS backend at its `mod` declaration is also strictly
stronger than the item-level gate it replaces: an `extern` added to
`rtems.rs` cannot outrun the C, because the whole file is behind
`all(target_os = "rtems", rtems_boot_linked)`.

Two guard tests follow the structure: the extern-scope test now asserts
the `mod` line, and a new one counts every `#[cfg(` in the funnel and
requires it to be a backend selection.

Pure move; no behavior change on any target.
